Remove unrefactored old code
